+++ src/queue/utils/constants.ts
@@ -1,6 +1,6 @@
 import { StoreDestinationArray } from "../../common/persistence/IExtentStore";
 
-export const VERSION = "3.20.0";
+export const VERSION = "3.20.1";
 export const QUEUE_API_VERSION = "2021-10-04";
 export const DEFAULT_QUEUE_SERVER_HOST_NAME = "127.0.0.1"; // Change to 0.0.0.0 when needs external access
 export const DEFAULT_QUEUE_LISTENING_PORT = 10001;